requirements.txt
# FastAPI Dashboard - Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
jinja2>=3.1.0
passlib>=1.7.4
cryptography>=41.0.0
